Revert "updating WPML"
This commit from the fw-v2 branch happened much later in May 2023 on the
"main" branch, thus reverting this commit to use the new commit when
merging "main".

This reverts commit 077fbef67f0ec91f062659cd322877aec6edd947.

+++ site/assets/plugins/sitepress-multilingual-cms/classes/ATE/Review/Cancel.php
@@ -2,10 +2,8 @@
 
 namespace WPML\TM\ATE\Review;
 
-use WPML\Ajax\IHandler;
-use WPML\Collect\Support\Collection;
-use WPML\FP\Cast;
 use WPML\FP\Fns;
+use WPML\FP\Logic;
 use WPML\FP\Obj;
 use WPML\FP\Relation;
 use WPML\LIB\WP\Post;
@@ -13,12 +11,9 @@ use WPML\TM\API\Job\Map;
 use WPML\TM\API\Jobs;
 use function WPML\FP\pipe;
 
-class Cancel implements IHandler {
+class Cancel {
 
-	public function run( Collection $data ) {
-		$jobIds       = $data->get( 'jobsIds' );
-		$deleteDrafts = $data->get( 'deleteDrafts' );
-
+	public static function run( array $jobIds ) {
 		$reviewJobs = wpml_collect( $jobIds )
 			->map( Jobs::get() )
 			->filter( ReviewStatus::doesJobNeedReview() );
@@ -27,40 +22,29 @@ class Cancel implements IHandler {
 			$reviewJobs->map( Obj::prop( 'job_id' ) )
 			           ->map( Jobs::clearReviewStatus() );
 
-			if ( $deleteDrafts ) {
-				$this->deleteDrafts( $reviewJobs );
-			}
+			if ( Relation::propEq( 'delete-drafts', "1", $_POST ) ) {
+				$doCancelJobAction = function ( $jobId ) {
+					$jobEntity = wpml_tm_get_jobs_repository()->get_job( Map::fromJobId( $jobId ), \WPML_TM_Job_Entity::POST_TYPE );
+					do_action( 'wpml_tm_job_cancelled', $jobEntity );
 
-			return $reviewJobs->pluck( 'job_id' )->map( Cast::toInt() );
+					return $jobId;
+				};
+
+				$getTranslatedId = Fns::memorize( pipe( Jobs::get(), Jobs::getTranslatedPostId() ) );
+
+				$isDraft = pipe( $getTranslatedId, Post::getStatus(), Relation::equals( 'draft' ) );
+
+				$deleteTranslatedPost = pipe( $getTranslatedId, Post::delete() );
+
+				$reviewJobs->map( Obj::prop( 'job_id' ) )
+				           ->map( Jobs::setNotTranslatedStatus() )
+				           ->map( Jobs::clearTranslated() )
+				           ->map( $doCancelJobAction )
+				           ->filter( $isDraft )
+				           ->map( Fns::tap( $deleteTranslatedPost ) )
+				           ->map( Jobs::delete() );
+			}
 		}
 
-		return [];
-	}
-
-	private function deleteDrafts( Collection $reviewJobs ) {
-		$doCancelJobsAction = function ( Collection $jobIds ) {
-			$getJobEntity = function ( $jobId ) {
-				return wpml_tm_get_jobs_repository()->get_job( Map::fromJobId( $jobId ), \WPML_TM_Job_Entity::POST_TYPE );
-			};
-
-			$jobEntities = $jobIds->map( $getJobEntity )->toArray();
-			do_action( 'wpml_tm_jobs_cancelled', $jobEntities );
-		};
-
-		$getTranslatedId = Fns::memorize( pipe( Jobs::get(), Jobs::getTranslatedPostId() ) );
-
-		$isDraft = pipe( $getTranslatedId, Post::getStatus(), Relation::equals( 'draft' ) );
-
-		$deleteTranslatedPost = pipe( $getTranslatedId, Post::delete() );
-
-		$reviewJobs = $reviewJobs->map( Obj::prop( 'job_id' ) )
-		                         ->map( Jobs::setNotTranslatedStatus() )
-		                         ->map( Jobs::clearTranslated() );
-
-		$doCancelJobsAction( $reviewJobs );
-
-		$reviewJobs->filter( $isDraft )
-		           ->map( Fns::tap( $deleteTranslatedPost ) )
-		           ->map( Jobs::delete() );
 	}
 }
